+++ jenkins_integration/artifacts/artifact_processor.py
@@ -18,10 +18,10 @@ workspace_root = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath
 if workspace_root not in sys.path:
     sys.path.insert(0, workspace_root)
 
-import jenkins-integration.config as config
-from jenkins-integration.github.github_workflow import _make_github_api_request
-from jenkins-integration.artifacts.ubai_client import upload_to_ubai
-from jenkins-integration.artifacts.artifactory_client import upload_to_artifactory
+import jenkins_integration.config as config
+from jenkins_integration.github.github_workflow import _make_github_api_request
+from jenkins_integration.artifacts.ubai_client import upload_to_ubai
+from jenkins_integration.artifacts.artifactory_client import upload_to_artifactory
 
 
 def download_and_upload_artifacts(workflow_id, branch_name, run_number, sqa=False):
