+++ jenkins_integration/artifacts/artifactory_client.py
@@ -17,7 +17,7 @@ workspace_root = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath
 if workspace_root not in sys.path:
     sys.path.insert(0, workspace_root)
 
-import jenkins-integration.config as config
+import jenkins_integration.config as config
 
 
 def upload_to_artifactory(filepath, artifact_name, branch_name, build_number):
