+++ jenkins_integration/github/github_workflow.py
@@ -18,7 +18,7 @@ workspace_root = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath
 if workspace_root not in sys.path:
     sys.path.insert(0, workspace_root)
 
-import jenkins-integration.config as config
+import jenkins_integration.config as config
 
 
 def get_latest_sha(branch_name):
